version: move to new package: cmd/wksctl/version
